backend/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
python-multipart==0.0.9
pydantic==2.8.0
pydantic-settings==2.2.0
httpx==0.27.0
itsdangerous==2.1.2
google-api-python-client==2.100.0
google-auth==2.23.3
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-cloud-firestore==2.12.0
python-dotenv==1.0.0
PyPDF2==3.0.1
python-pptx==0.6.21
pytest==7.4.2
pytest-asyncio==0.21.1
# Linting and formatting
flake8==6.1.0
black==23.12.0
mypy==1.8.0
isort==5.13.2
# Logging
structlog==23.2.0
# Pre-commit
pre-commit==3.5.0
